move FileUpload extended functionality from project-base to the framework package
