+++ project-base/app/config/services.yaml
@@ -37,9 +37,6 @@ services:
         tags:
             - { name: doctrine.event_subscriber, priority: -100 }
 
-    App\Component\FileUpload\FileUpload:
-        arguments: ['%shopsys.var_dir%', '%shopsys.uploaded_file_dir%', '%shopsys.image_dir%']
-
     App\Component\Image\ImageFacade:
         arguments:
             $imageUrlPrefix: '%shopsys.image_url_prefix%'
@@ -197,9 +194,6 @@ services:
 
     Shopsys\FrameworkBundle\Component\Domain\Multidomain\MultidomainEntityClassProviderInterface:
         alias: App\Model\MultidomainEntityClassProvider
-
-    Shopsys\FrameworkBundle\Component\FileUpload\FileUpload:
-        alias: App\Component\FileUpload\FileUpload
 
     Shopsys\FrameworkBundle\Component\Image\ImageFacade:
         alias: App\Component\Image\ImageFacade
